Move coords to fontdrasil so fea-rs can agree with us on what NormalizedLocation looks like
