+++ fontbe/src/os2.rs
@@ -851,13 +851,10 @@ impl Work<Context, AnyWorkId, Error> for Os2Work {
 
 #[cfg(test)]
 mod tests {
-    use std::collections::HashSet;
-
-    use fontir::{
-        coords::NormalizedLocation,
-        ir::{GlobalMetric, GlobalMetrics},
-    };
+    use fontdrasil::coords::NormalizedLocation;
+    use fontir::ir::{GlobalMetric, GlobalMetrics};
     use read_fonts::types::Tag;
+    use std::collections::HashSet;
     use write_fonts::tables::os2::Os2;
 
     use crate::os2::codepage_range_bits;
