+++ fontbe/src/test_util.rs
@@ -1,9 +1,9 @@
 //! Only included in test
 
 use font_types::Tag;
-use fontir::{
+use fontdrasil::{
     coords::{CoordConverter, DesignCoord, UserCoord},
-    ir::Axis,
+    types::Axis,
 };
 use std::cmp;
