+++ fontdrasil/src/lib.rs
@@ -1,5 +1,8 @@
 //! Helper library for code that is of value to FE and BE of font compilation.
 
+pub mod coords;
 pub mod orchestration;
 pub mod paths;
+pub mod piecewise_linear_map;
+mod serde;
 pub mod types;
