+++ fontdrasil/src/types.rs
@@ -7,6 +7,10 @@ use std::fmt::{Debug, Display};
 use serde::{Deserialize, Serialize};
 use smol_str::SmolStr;
 
+use font_types::Tag;
+
+use crate::coords::{CoordConverter, UserCoord};
+
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
 pub struct GlyphName(SmolStr);
 
@@ -49,3 +53,20 @@ impl Display for GlyphName {
 
 pub type GroupName = GlyphName;
 pub type AnchorName = GlyphName;
+
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
+pub struct Axis {
+    pub name: String,
+    pub tag: Tag,
+    pub min: UserCoord,
+    pub default: UserCoord,
+    pub max: UserCoord,
+    pub hidden: bool,
+    pub converter: CoordConverter,
+}
+
+impl Axis {
+    pub fn is_point(&self) -> bool {
+        self.min == self.default && self.max == self.default
+    }
+}
