+++ fontir/src/lib.rs
@@ -1,10 +1,8 @@
-pub mod coords;
 pub mod error;
 pub mod glyph;
 pub mod ir;
 pub mod orchestration;
 pub mod paths;
-pub mod piecewise_linear_map;
 pub(crate) mod serde;
 pub mod source;
 pub mod stateset;
